requirements.txt
cfgv==3.4.0
distlib==0.3.8
filelock==3.13.1
identify==2.5.33
iniconfig==2.0.0
isort==5.13.2
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pytest==8.0.0
PyYAML==6.0.1
types-psycopg2==2.9.21.20240201
typing_extensions==4.9.0
virtualenv==20.25.0
